Split shell run enqueue tests
